pyproject.toml
[project]
name = "styled-studio"
version = "0.0.1"
description = "Advanced styling and custom components for Streamlit applications"
authors = [
    {name = "JonathanAlles", email = "Jonathan.Alles@evo-byte.com"}
]
readme = "readme.md"
requires-python = ">=3.10"  # Changed from >=3.11 for broader compatibility
license = {text = "MIT"}
keywords = ["streamlit", "styling", "css", "components", "ui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: User Interfaces"
]
dependencies = [
    "streamlit>=1.42.0",  # Add Streamlit peer dependency (minimum version 1.42)
    "pandas>=2.0.0,<3.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "st-styled @ file:///home/jona/scribble/st-styled/dist/st_styled-0.0.3-py3-none-any.whl",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0"
]

[project.urls]
Homepage = "https://st-styled.evo-byte.com/"  # Note: project repo uses hyphen
Repository = "https://github.com/EvobyteDigitalBiology/st-styled"  # Note: project repo uses hyphen
Issues = "https://github.com/EvobyteDigitalBiology/st-styled/issues"
Documentation = "https://st-styled.evo-byte.com/"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
ruff = "^0.1.0"
mypy = "^1.5.0"
pre-commit = "^3.0.0"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "COM812", "ISC001", "A001", "A002", "SIM118", "I001", "Q000", "UP", "RUF012", "FBT", "TRY300", "BLE001", "RUF010", "PLR0911", "RET505", "B028", "PGH003", "SIM108", "PLW2901"]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101"]  # Allow assert in tests

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=st_yled --cov-report=term-missing"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["st_yled"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
